Refactor integration test (part 2) (#627)
