--- test/integration/eslint-config-unicorn-tester/package.json
@@ -1,11 +1,0 @@
-{
-	"name": "eslint-config-unicorn-tester",
-	"dependencies": {
-		"@typescript-eslint/parser": "2.24.0",
-		"babel-eslint": "^10.1.0",
-		"eslint": "^6.8.0",
-		"eslint-plugin-unicorn": "file:../../..",
-		"typescript": "3.8.3",
-		"vue-eslint-parser": "7.0.0"
-	}
-}
