+++ test/integration/package.json
@@ -1,0 +1,7 @@
+{
+	"private": true,
+	"name": "eslint-config-unicorn-tester",
+	"dependencies": {
+		"eslint-plugin-unicorn": "file:../.."
+	}
+}
