+++ test/integration/test.js
@@ -6,135 +6,7 @@ const execa = require('execa');
 const del = require('del');
 const chalk = require('chalk');
 const {isCI} = require('ci-info');
-
-const typescriptArguments = ['--parser', '@typescript-eslint/parser', '--ext', '.ts'];
-const vueArguments = ['--parser', 'vue-eslint-parser', '--ext', '.vue'];
-
-const projects = [
-	'https://github.com/avajs/ava',
-	'https://github.com/chalk/chalk',
-	'https://github.com/chalk/wrap-ansi',
-	'https://github.com/sindresorhus/np',
-	'https://github.com/sindresorhus/ora',
-	'https://github.com/sindresorhus/p-map',
-	'https://github.com/sindresorhus/os-locale',
-	'https://github.com/sindresorhus/execa',
-	'https://github.com/sindresorhus/pify',
-	'https://github.com/sindresorhus/boxen',
-	'https://github.com/sindresorhus/make-dir',
-	'http://github.com/SamVerschueren/listr',
-	'http://github.com/SamVerschueren/listr-update-renderer',
-	'http://github.com/SamVerschueren/clinton',
-	'http://github.com/SamVerschueren/bragg',
-	'http://github.com/SamVerschueren/bragg-router',
-	'http://github.com/SamVerschueren/dev-time',
-	'https://github.com/SamVerschueren/decode-uri-component',
-	'https://github.com/kevva/to-ico',
-	'https://github.com/kevva/download',
-	'https://github.com/kevva/brightness',
-	'https://github.com/kevva/decompress',
-	'https://github.com/kevva/npm-conf',
-	'https://github.com/imagemin/imagemin',
-	'https://github.com/qix-/color-convert',
-	'https://github.com/sindresorhus/ky',
-	'https://github.com/sindresorhus/query-string',
-	'https://github.com/sindresorhus/meow',
-	'https://github.com/sindresorhus/globby',
-	'https://github.com/sindresorhus/emittery',
-	{
-		repository: 'https://github.com/sindresorhus/p-queue',
-		extraArguments: typescriptArguments
-	},
-	'https://github.com/sindresorhus/pretty-bytes',
-	'https://github.com/sindresorhus/normalize-url',
-	{
-		repository: 'https://github.com/sindresorhus/pageres',
-		extraArguments: typescriptArguments
-	},
-	{
-		repository: 'https://github.com/sindresorhus/got',
-		extraArguments: typescriptArguments
-	},
-	// TODO: Add this project when #561 got fixed
-	// {
-	// 	repository: 'https://github.com/eslint/eslint',
-	// 	path: 'lib'
-	// },
-	{
-		repository: 'https://github.com/prettier/prettier',
-		path: 'src'
-	},
-	{
-		repository: 'https://github.com/facebook/react',
-		path: 'packages'
-	},
-	{
-		repository: 'https://github.com/angular/angular',
-		path: 'packages',
-		extraArguments: typescriptArguments
-	},
-	{
-		repository: 'https://github.com/microsoft/typescript',
-		path: 'src',
-		extraArguments: typescriptArguments
-	},
-	// TODO: Add this project when `@typescript-eslint/parser` support `Type-Only Imports and Export`
-	// {
-	// 	repository: 'https://github.com/microsoft/vscode',
-	// 	path: 'src/vs',
-	// 	extraArguments: typescriptArguments
-	// },
-	{
-		repository: 'https://github.com/ElemeFE/element',
-		path: 'packages',
-		extraArguments: vueArguments
-	},
-	{
-		repository: 'https://github.com/iview/iview',
-		path: 'src',
-		extraArguments: vueArguments
-	},
-	'https://github.com/sindresorhus/create-dmg',
-	'https://github.com/sindresorhus/cp-file',
-	'https://github.com/sindresorhus/capture-website',
-	'https://github.com/sindresorhus/file-type',
-	'https://github.com/sindresorhus/slugify',
-	'https://github.com/gatsbyjs/gatsby',
-	{
-		repository: 'https://github.com/puppeteer/puppeteer',
-		path: 'lib'
-	},
-	{
-		repository: 'https://github.com/zeit/next.js',
-		path: 'packages',
-		extraArguments: typescriptArguments
-	},
-	'https://github.com/chakra-ui/chakra-ui',
-	'https://github.com/ReactTraining/react-router',
-	'https://github.com/facebook/relay',
-	'https://github.com/mozilla/pdf.js'
-].map(project => {
-	if (typeof project === 'string') {
-		project = {repository: project};
-	}
-
-	const {
-		repository,
-		name = repository.split('/').pop(),
-		path = '',
-		extraArguments = []
-	} = project;
-
-	return {
-		...project,
-		name,
-		repository,
-		path,
-		extraArguments
-	};
-});
-
-const cwd = path.join(__dirname, 'eslint-config-unicorn-tester');
+const projects = require('./projects');
 
 const enrichErrors = (packageName, cliArguments, f) => async (...arguments_) => {
 	try {
@@ -146,19 +18,19 @@ const enrichErrors = (packageName, cliArguments, f) => async (...arguments_) => 
 	}
 };
 
-const makeEslintTask = (project, destination, extraArguments = []) => {
+const makeEslintTask = (project, destination) => {
 	const arguments_ = [
 		'eslint',
+		'--fix-dry-run',
 		'--no-eslintrc',
 		// https://github.com/microsoft/TypeScript/blob/a1c8608f681488fda3f0b6ffd89195a81f80f0b0/.eslintignore#L6
 		project.name === 'typescript' ? '' : '--no-ignore',
 		'--format',
 		'json',
 		'--config',
-		path.join(cwd, 'index.js'),
+		path.join(__dirname, 'config.js'),
 		project.path ? path.join(destination, project.path) : destination,
-		...project.extraArguments,
-		...extraArguments
+		...project.extraArguments
 	].filter(Boolean);
 
 	return enrichErrors(project.name, arguments_, async () => {
@@ -202,16 +74,23 @@ const makeEslintTask = (project, destination, extraArguments = []) => {
 };
 
 const execute = project => {
-	const destination = path.join(cwd, 'fixtures', project.name);
+	const destination = path.join(__dirname, 'fixtures', project.name);
 
 	return new Listr([
 		{
 			title: 'Cloning',
-			task: () => execa('git', ['clone', project.repository, '--single-branch', destination])
+			task: () => execa('git', [
+				'clone',
+				project.repository,
+				'--single-branch',
+				'--depth',
+				'1',
+				destination
+			])
 		},
 		{
 			title: 'Running eslint',
-			task: makeEslintTask(project, destination, ['--fix-dry-run'])
+			task: makeEslintTask(project, destination)
 		},
 		{
 			title: 'Clean up',
@@ -228,7 +107,7 @@ const execute = project => {
 const list = new Listr([
 	{
 		title: 'Setup',
-		task: () => execa('npm', ['install'], {cwd})
+		task: () => execa('npm', ['install'], {cwd: __dirname})
 	},
 	{
 		title: 'Integration tests',
